requirements.txt
Django>=3.2.4,<3.3
Djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow>=8.2.0,<8.3

cryptography~=2.8
django-cors-headers~=3.1.1
drf-jwt~=1.13.3
requests~=2.22.0
pyjwt~=1.7.1
